requirements.txt
django>=5.0.0
djangorestframework>=3.15.0
python-decouple>=3.8
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
stripe >= 12.3.0
channels==4.2.2
channels-redis==4.2.1
redis==6.2.0
daphne>=4.2.1
reportlab>=4.4.0
PyJWT>=2.8.0
